Initial commit. Basic JIT repl with interpreter mode
No optimization, grammar is limited to mathematical expressions.

Also includes some matrix code which will eventually be part of
the runtime, but not yet.

Also includes some nanojit support, likely never to be used, but
keeping my options open.

AST is templated which I am still trying to decide if that is a
mistake or not...Likely to be removed, especially since I should
really be converting it to SSA, meaning the AST itself likely
doesn't need any resolution (its not strongly typed, etc)

Main file just provides repl support and a no-iteration loop used
for basic profiling.