Clear varsToPhi before an if statement's true blocks.
Suggested by a blog post reader -- thank you tinkerzf!

I'm not positive this is correct, since I haven't worked with SSA in a
,hile, that this is correct. However, varsToPhi is being cleared before
the false statements, so either that's broken or this line is indeed
missing.

I think its that this line is missing -- phi'ing too many vars should be
safe, and removed in optimizations/generate harmless code later. I don't
think there's a reason to phi vars changed in an outer block:

if (x) {
  x = true;
  if (y) {
    // no changes to x
  }
  // x shouldn't be phi'd here!
}
// x should only be phi'd here.

Adding this did not break any tests (not that my coverage is that
great). And I didn't add any since I don't think it would affect the
runtime, only the performance.